util: Combine and return all data on any image.
This commit makes lots of strides in terms of what is possible
to do with Scarfin. Now, the mounting maps can make use of tags,
devices and pre-parsed date taken data, allowing useful mounts
like tag based folders or album/tag heirarchy.

We had to switch to snake case for our database fields and this
is a breaking change so we cannot migrate from past state of the
database to current. We're in alpha; it isn't worth it trying to
make a migration for this one.

We write some more about what read_paths flag in config does
and make sure Scarfin behaves the same with both true/false
values for read_paths.